refactor: move working folder into src,  add measurements and pauli matrices
