+++ src/tncdr/stabilizer_mps/pauli_string.py
