+++ src/tncdr/stabilizer_mps/tensor_network.py
@@ -2,13 +2,18 @@ import networkx as nx
 import numpy as np
 
 from typing import Union
+from tncdr.stabilizer_mps.tn_utils import paulis
 
 from dataclasses import dataclass
 
 @dataclass
 class TensorNetwork:
 
-    #TODO Add measurements
+    #TODO Add key tensors:
+    # - Measurement: shape -> (2,)
+    # - Copy: shape -> (n,n,n), n given in input
+    # - Pauli: shape -> (2,2)
+    # - PauliRot: shape -> (2,2,2)
 
     def __post_init__(self):
         self.tensornet = nx.MultiDiGraph()
@@ -19,6 +24,19 @@ class TensorNetwork:
             tensor=tensor, 
             shape=tensor.shape, 
             free_directions=[True]*len(tensor.shape)
+        )
+
+    def add_measurement(self, id:str, alpha:float=1.0, beta:float=0.0):
+        self.add_tensor(
+            id=id,
+            tensor=np.array([alpha, beta])
+        )
+
+    def add_pauli_pair(self, p0:str, p1:str):
+        tensor = np.array([paulis[p0], paulis[p1]])
+        self.add_tensor(
+            id=f'{p0}{p1} pair',
+            tensor=tensor,
         )
 
     def add_edge(self, node_in:str, node_out:str, edge_id:str, directions:tuple[int, int], **edge_metadata):
@@ -117,10 +135,10 @@ class TensorNetwork:
         # Remove the edges from the graph
         for edge_id in edge_ids:
             self.remove_edge(node_in=node_in, node_out=node_out, edge_id=edge_id)
-
+            
         # Add new node, containing contracted tensors
         self.add_tensor(id=new_node_id, tensor=new_tensor)
-
+        
         # Transfer the edge connections from the old to the new node and delete it
         self._reconnect_edges(
             node=node_in, 
@@ -134,6 +152,9 @@ class TensorNetwork:
             shift=len(non_contracted_index_in), # Comply with indexing convention of numpy tensordot
         )
     
+    def _partial_trace(self, *args):
+        pass
+
     def _reconnect_edges(self, node:str, new_node_id:str, survived_directions:list, shift:int=0):
 
         # First we take all the edges entering the node
@@ -143,7 +164,7 @@ class TensorNetwork:
                 # Updated tensor direction in the new node corresponding to the edge
                 survived_directions.index(metadata['directions'][0]) + shift,
                 # Kept direction on the connected node
-                metadata['directions'][1], 
+                metadata['directions'][1],
             )
 
             self.remove_edge(node_in=u, node_out=v, edge_id=edge_id)
@@ -151,7 +172,7 @@ class TensorNetwork:
         
         # Second we take all the edges exiting the node
         for u,v,edge_id,metadata in list(self.tensornet.in_edges(nbunch=node, keys=True, data=True)):
-            
+
             directions = (
                 # Kept direction on the connected node
                 metadata['directions'][0],
@@ -163,4 +184,4 @@ class TensorNetwork:
 
         # Remove node
         self.tensornet.remove_node(node)
-    
+
