+++ test_pauli.py
